ai_service/requirements.txt
fastapi~=0.115.6
uvicorn
pydantic~=2.10.5
numpy~=2.2.1
pandas~=2.2.3
tqdm~=4.67.1
python-multipart
torch~=2.5.1
faiss-cpu
Pillow~=11.1.0
matplotlib~=3.10.0
ultralytics~=8.3.59
transformers
git+https://github.com/openai/CLIP.git
openpyxl

torchvision~=0.20.1